refactor: 로그아웃 버튼을 헤더에서 마이페이지로 이동
