+++ loneleap-client/src/components/Header.jsx
@@ -1,25 +1,9 @@
 import React from "react";
-import { Link, useNavigate } from "react-router-dom";
-import { useSelector, useDispatch } from "react-redux";
-
-import { logout } from "services/auth";
-import { clearUser } from "store/userSlice";
+import { Link } from "react-router-dom";
+import { useSelector } from "react-redux";
 
 export default function Header() {
   const user = useSelector((state) => state.user.user);
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
-
-  const handleLogout = async () => {
-    try {
-      await logout(); // Firebase 로그아웃
-      dispatch(clearUser());
-      navigate("/");
-    } catch (error) {
-      console.error("로그아웃 실패:", error);
-      alert("로그아웃에 실패했습니다. 다시 시도해주세요.");
-    }
-  };
 
   return (
     <header className="bg-white px-6 py-4 border-b flex justify-between items-center shadow-sm">
@@ -42,12 +26,6 @@ export default function Header() {
           <Link to="/mypage" className="hover:text-black">
             마이페이지
           </Link>
-          <button
-            onClick={handleLogout}
-            className="text-red-500 hover:underline ml-2"
-          >
-            로그아웃
-          </button>
         </div>
       ) : (
         <div className="flex gap-4 text-sm text-gray-700">
